+++ ListableUI/Sources/ListScrollPositionInfo.swift
@@ -29,6 +29,15 @@ public struct ListScrollPositionInfo : Equatable {
     
     /// If the last item list is partially visible.
     public var isLastItemVisible : Bool
+
+    /// Distance required to scroll to the bottom
+    public var bottomScrollOffset: CGFloat
+
+    /// `bounds` of the list view
+    public var bounds: CGRect
+
+    /// `safeAreaInsests` of the list view
+    public var safeAreaInsets: UIEdgeInsets
     
     ///
     /// Used to retrieve the visible content edges for the list's content.
@@ -106,6 +115,11 @@ public struct ListScrollPositionInfo : Equatable {
         
         self.isFirstItemVisible = isFirstItemVisible
         self.isLastItemVisible = isLastItemVisible
+
+        self.bottomScrollOffset = scrollView.contentSize.height - scrollView.bounds.size.height - scrollView.contentOffset.y + scrollView.adjustedContentInset.bottom
+
+        self.bounds = scrollView.bounds
+        self.safeAreaInsets = scrollView.safeAreaInsets
     }
     
     struct ScrollViewState : Equatable