+++ ListableUI/Sources/Internal/PresentationState/PresentationState.SectionState.swift
@@ -67,6 +67,7 @@ extension PresentationState
             new newSection : Section,
             changes : SectionedDiff<Section, AnyIdentifier, AnyItem, AnyIdentifier>.ItemChanges,
             reason: ApplyReason,
+            animated : Bool,
             dependencies : ItemStateDependencies,
             updateCallbacks : UpdateCallbacks
         ) {
@@ -82,6 +83,7 @@ extension PresentationState
                 ),
                 new: self.model.header,
                 reason: reason,
+                animated: animated,
                 updateCallbacks: updateCallbacks,
                 environment: environment
             )
@@ -94,6 +96,7 @@ extension PresentationState
                 ),
                 new: self.model.footer,
                 reason: reason,
+                animated: animated,
                 updateCallbacks: updateCallbacks,
                 environment: environment
             )