+++ ListableUI/Sources/Internal/PresentationState/PresentationState.swift
@@ -277,6 +277,7 @@ final class PresentationState
         with diff : SectionedDiff<Section, AnyIdentifier, AnyItem, AnyIdentifier>,
         slice : Content.Slice,
         reason: ApplyReason,
+        animated : Bool,
         dependencies: ItemStateDependencies,
         updateCallbacks : UpdateCallbacks,
         loggable : SignpostLoggable?
@@ -301,6 +302,7 @@ final class PresentationState
             ),
             new: slice.content.containerHeader,
             reason: reason,
+            animated: animated,
             updateCallbacks: updateCallbacks,
             environment: environment
         )
@@ -313,6 +315,7 @@ final class PresentationState
             ),
             new: slice.content.header,
             reason: reason,
+            animated: animated,
             updateCallbacks: updateCallbacks,
             environment: environment
         )
@@ -325,6 +328,7 @@ final class PresentationState
             ),
             new: slice.content.footer,
             reason: reason,
+            animated: animated,
             updateCallbacks: updateCallbacks,
             environment: environment
         )
@@ -337,6 +341,7 @@ final class PresentationState
             ),
             new: slice.content.overscrollFooter,
             reason: reason,
+            animated: animated,
             updateCallbacks: updateCallbacks,
             environment: environment
         )
@@ -360,6 +365,7 @@ final class PresentationState
                     new: new,
                     changes: changes,
                     reason: reason,
+                    animated: animated,
                     dependencies: dependencies,
                     updateCallbacks: updateCallbacks
                 )
@@ -370,6 +376,7 @@ final class PresentationState
                     new: new,
                     changes: changes,
                     reason: reason,
+                    animated: animated,
                     dependencies: dependencies,
                     updateCallbacks: updateCallbacks
                 )