+++ ListableUI/Sources/ListView/ListView.swift
@@ -987,6 +987,7 @@ public final class ListView : UIView, KeyboardObserverDelegate
                 with: diff,
                 slice: visibleSlice,
                 reason: .wasUpdated,
+                animated: reason.animated,
                 dependencies: dependencies,
                 updateCallbacks: updateCallbacks,
                 loggable: self